+++ src/form/wr.jl
@@ -50,9 +50,8 @@ function constraint_kcl_shunt_ne{T <: AbstractWRForm}(pm::GenericPowerModel{T}, 
     pg = pm.var[:pg]
     qg = pm.var[:qg]
 
-    c1 = @constraint(pm.model, sum(p[a] for a in bus_arcs) + sum(p_ne[a] for a in bus_arcs_ne) + sum(p_dc[a_dc] for a_dc in bus_arcs_dc) == sum(pg[g] for g in bus_gens) - pd - gs*w)
-    c2 = @constraint(pm.model, sum(q[a] for a in bus_arcs) + sum(q_ne[a] for a in bus_arcs_ne) + sum(p_dc[a_dc] for a_dc in bus_arcs_dc) == sum(qg[g] for g in bus_gens) - qd + bs*w)
-    return Set([c1, c2])
+    @constraint(pm.model, sum(p[a] for a in bus_arcs) + sum(p_ne[a] for a in bus_arcs_ne) + sum(p_dc[a_dc] for a_dc in bus_arcs_dc) == sum(pg[g] for g in bus_gens) - pd - gs*w)
+    @constraint(pm.model, sum(q[a] for a in bus_arcs) + sum(q_ne[a] for a in bus_arcs_ne) + sum(p_dc[a_dc] for a_dc in bus_arcs_dc) == sum(qg[g] for g in bus_gens) - qd + bs*w)
 end
 
 
@@ -72,9 +71,8 @@ function constraint_ohms_yt_from_ne{T <: AbstractWRForm}(pm::GenericPowerModel{T
     wr = pm.var[:wr_ne][i]
     wi = pm.var[:wi_ne][i]
 
-    c1 = @constraint(pm.model, p_fr == g/tm*w_fr + (-g*tr+b*ti)/tm*(wr) + (-b*tr-g*ti)/tm*( wi) )
-    c2 = @constraint(pm.model, q_fr == -(b+c/2)/tm*w_fr - (-b*tr-g*ti)/tm*(wr) + (-g*tr+b*ti)/tm*( wi) )
-    return Set([c1, c2])
+    @constraint(pm.model, p_fr == g/tm*w_fr + (-g*tr+b*ti)/tm*(wr) + (-b*tr-g*ti)/tm*( wi) )
+    @constraint(pm.model, q_fr == -(b+c/2)/tm*w_fr - (-b*tr-g*ti)/tm*(wr) + (-g*tr+b*ti)/tm*( wi) )
 end
 
 """
@@ -92,9 +90,8 @@ function constraint_ohms_yt_to_ne{T <: AbstractWRForm}(pm::GenericPowerModel{T},
     wr = pm.var[:wr_ne][i]
     wi = pm.var[:wi_ne][i]
 
-    c1 = @constraint(pm.model, p_to == g*w_to + (-g*tr-b*ti)/tm*(wr) + (-b*tr+g*ti)/tm*(-wi) )
-    c2 = @constraint(pm.model, q_to == -(b+c/2)*w_to - (-b*tr+g*ti)/tm*(wr) + (-g*tr-b*ti)/tm*(-wi) )
-    return Set([c1, c2])
+    @constraint(pm.model, p_to == g*w_to + (-g*tr-b*ti)/tm*(wr) + (-b*tr+g*ti)/tm*(-wi) )
+    @constraint(pm.model, q_to == -(b+c/2)*w_to - (-b*tr+g*ti)/tm*(wr) + (-g*tr-b*ti)/tm*(-wi) )
 end
 
 ""
@@ -104,11 +101,9 @@ function constraint_phase_angle_difference{T <: AbstractWRForm}(pm::GenericPower
     wr = pm.var[:wr][(f_bus, t_bus)]
     wi = pm.var[:wi][(f_bus, t_bus)]
 
-    c1 = @constraint(pm.model, wi <= tan(angmax)*wr)
-    c2 = @constraint(pm.model, wi >= tan(angmin)*wr)
-    c3 = cut_complex_product_and_angle_difference(pm.model, w_fr, w_to, wr, wi, angmin, angmax)
-
-    return Set([c1, c2, c3])
+    @constraint(pm.model, wi <= tan(angmax)*wr)
+    @constraint(pm.model, wi >= tan(angmin)*wr)
+    cut_complex_product_and_angle_difference(pm.model, w_fr, w_to, wr, wi, angmin, angmax)
 end
 
 ""
@@ -137,20 +132,15 @@ function constraint_voltage_on_off{T <: AbstractWRForm}(pm::GenericPowerModel{T}
     w_from = pm.var[:w_from]
     w_to = pm.var[:w_to]
 
-    cs = Set()
-    cs1 = constraint_voltage_magnitude_sqr_from_on_off(pm)
-    cs2 = constraint_voltage_magnitude_sqr_to_on_off(pm)
-    cs3 = constraint_voltage_product_on_off(pm)
-    cs = union(cs, cs1, cs2, cs3)
+    constraint_voltage_magnitude_sqr_from_on_off(pm)
+    constraint_voltage_magnitude_sqr_to_on_off(pm)
+    constraint_voltage_product_on_off(pm)
 
     for (l,i,j) in pm.ref[:arcs_from]
-        cs4 = relaxation_complex_product_on_off(pm.model, w[i], w[j], wr[l], wi[l], z[l])
-        cs5 = relaxation_equality_on_off(pm.model, w[i], w_from[l], z[l])
-        cs6 = relaxation_equality_on_off(pm.model, w[j], w_to[l], z[l])
-        cs = union(cs, cs4, cs5, cs6)
+        relaxation_complex_product_on_off(pm.model, w[i], w[j], wr[l], wi[l], z[l])
+        relaxation_equality_on_off(pm.model, w[i], w_from[l], z[l])
+        relaxation_equality_on_off(pm.model, w[j], w_to[l], z[l])
     end
-
-    return cs
 end
 
 ""
@@ -169,25 +159,22 @@ function constraint_voltage_ne{T <: AbstractWRForm}(pm::GenericPowerModel{T})
     w_from = pm.var[:w_from_ne]
     w_to = pm.var[:w_to_ne]
 
-    cs = Set()
     for (l,i,j) in pm.ref[:ne_arcs_from]
-        c1 = @constraint(pm.model, w_from[l] <= z[l]*buses[branches[l]["f_bus"]]["vmax"]^2)
-        c2 = @constraint(pm.model, w_from[l] >= z[l]*buses[branches[l]["f_bus"]]["vmin"]^2)
+        @constraint(pm.model, w_from[l] <= z[l]*buses[branches[l]["f_bus"]]["vmax"]^2)
+        @constraint(pm.model, w_from[l] >= z[l]*buses[branches[l]["f_bus"]]["vmin"]^2)
 
-        c3 = @constraint(pm.model, wr[l] <= z[l]*wr_max[bi_bp[l]])
-        c4 = @constraint(pm.model, wr[l] >= z[l]*wr_min[bi_bp[l]])
-        c5 = @constraint(pm.model, wi[l] <= z[l]*wi_max[bi_bp[l]])
-        c6 = @constraint(pm.model, wi[l] >= z[l]*wi_min[bi_bp[l]])
+        @constraint(pm.model, wr[l] <= z[l]*wr_max[bi_bp[l]])
+        @constraint(pm.model, wr[l] >= z[l]*wr_min[bi_bp[l]])
+        @constraint(pm.model, wi[l] <= z[l]*wi_max[bi_bp[l]])
+        @constraint(pm.model, wi[l] >= z[l]*wi_min[bi_bp[l]])
 
-        c7 = @constraint(pm.model, w_to[l] <= z[l]*buses[branches[l]["t_bus"]]["vmax"]^2)
-        c8 = @constraint(pm.model, w_to[l] >= z[l]*buses[branches[l]["t_bus"]]["vmin"]^2)
+        @constraint(pm.model, w_to[l] <= z[l]*buses[branches[l]["t_bus"]]["vmax"]^2)
+        @constraint(pm.model, w_to[l] >= z[l]*buses[branches[l]["t_bus"]]["vmin"]^2)
 
-        c9 = relaxation_complex_product_on_off(pm.model, w[i], w[j], wr[l], wi[l], z[l])
-        c10 = relaxation_equality_on_off(pm.model, w[i], w_from[l], z[l])
-        c11 = relaxation_equality_on_off(pm.model, w[j], w_to[l], z[l])
-        cs = Set([cs, c1, c2, c3, c4, c5, c6, c7, c8,c9, c10, c11])
+        relaxation_complex_product_on_off(pm.model, w[i], w[j], wr[l], wi[l], z[l])
+        relaxation_equality_on_off(pm.model, w[i], w_from[l], z[l])
+        relaxation_equality_on_off(pm.model, w[j], w_to[l], z[l])
     end
-    return cs
 end
 
 
@@ -199,14 +186,10 @@ function constraint_voltage_magnitude_from_on_off{T <: AbstractWRForm}(pm::Gener
     v_from = pm.var[:v_from]
     z = pm.var[:line_z]
 
-    cs = Set()
     for (i, branch) in pm.ref[:branch]
-        c1 = @constraint(pm.model, v_from[i] <= z[i]*buses[branch["f_bus"]]["vmax"])
-        c2 = @constraint(pm.model, v_from[i] >= z[i]*buses[branch["f_bus"]]["vmin"])
-        push!(cs, c1)
-        push!(cs, c2)
+        @constraint(pm.model, v_from[i] <= z[i]*buses[branch["f_bus"]]["vmax"])
+        @constraint(pm.model, v_from[i] >= z[i]*buses[branch["f_bus"]]["vmin"])
     end
-    return cs
 end
 
 ""
@@ -217,14 +200,10 @@ function constraint_voltage_magnitude_to_on_off{T <: AbstractWRForm}(pm::Generic
     v_to = pm.var[:v_to]
     z = pm.var[:line_z]
 
-    cs = Set()
     for (i, branch) in pm.ref[:branch]
-        c1 = @constraint(pm.model, v_to[i] <= z[i]*buses[branch["t_bus"]]["vmax"])
-        c2 = @constraint(pm.model, v_to[i] >= z[i]*buses[branch["t_bus"]]["vmin"])
-        push!(cs, c1)
-        push!(cs, c2)
+        @constraint(pm.model, v_to[i] <= z[i]*buses[branch["t_bus"]]["vmax"])
+        @constraint(pm.model, v_to[i] >= z[i]*buses[branch["t_bus"]]["vmin"])
     end
-    return cs
 end
 
 
@@ -236,14 +215,10 @@ function constraint_voltage_magnitude_sqr_from_on_off{T <: AbstractWRForm}(pm::G
     w_from = pm.var[:w_from]
     z = pm.var[:line_z]
 
-    cs = Set()
     for (i, branch) in pm.ref[:branch]
-        c1 = @constraint(pm.model, w_from[i] <= z[i]*buses[branch["f_bus"]]["vmax"]^2)
-        c2 = @constraint(pm.model, w_from[i] >= z[i]*buses[branch["f_bus"]]["vmin"]^2)
-        push!(cs, c1)
-        push!(cs, c2)
+        @constraint(pm.model, w_from[i] <= z[i]*buses[branch["f_bus"]]["vmax"]^2)
+        @constraint(pm.model, w_from[i] >= z[i]*buses[branch["f_bus"]]["vmin"]^2)
     end
-    return cs
 end
 
 ""
@@ -254,14 +229,10 @@ function constraint_voltage_magnitude_sqr_to_on_off{T <: AbstractWRForm}(pm::Gen
     w_to = pm.var[:w_to]
     z = pm.var[:line_z]
 
-    cs = Set()
     for (i, branch) in pm.ref[:branch]
-        c1 = @constraint(pm.model, w_to[i] <= z[i]*buses[branch["t_bus"]]["vmax"]^2)
-        c2 = @constraint(pm.model, w_to[i] >= z[i]*buses[branch["t_bus"]]["vmin"]^2)
-        push!(cs, c1)
-        push!(cs, c2)
+        @constraint(pm.model, w_to[i] <= z[i]*buses[branch["t_bus"]]["vmax"]^2)
+        @constraint(pm.model, w_to[i] >= z[i]*buses[branch["t_bus"]]["vmin"]^2)
     end
-    return cs
 end
 
 ""
@@ -274,18 +245,12 @@ function constraint_voltage_product_on_off{T <: AbstractWRForm}(pm::GenericPower
     wi = pm.var[:wi]
     z = pm.var[:line_z]
 
-    cs = Set()
     for b in keys(pm.ref[:branch])
-        c1 = @constraint(pm.model, wr[b] <= z[b]*wr_max[bi_bp[b]])
-        c2 = @constraint(pm.model, wr[b] >= z[b]*wr_min[bi_bp[b]])
-        c3 = @constraint(pm.model, wi[b] <= z[b]*wi_max[bi_bp[b]])
-        c4 = @constraint(pm.model, wi[b] >= z[b]*wi_min[bi_bp[b]])
-        push!(cs, c1)
-        push!(cs, c2)
-        push!(cs, c3)
-        push!(cs, c4)
+        @constraint(pm.model, wr[b] <= z[b]*wr_max[bi_bp[b]])
+        @constraint(pm.model, wr[b] >= z[b]*wr_min[bi_bp[b]])
+        @constraint(pm.model, wi[b] <= z[b]*wi_max[bi_bp[b]])
+        @constraint(pm.model, wi[b] >= z[b]*wi_min[bi_bp[b]])
     end
-    return cs
 end
 
 """
@@ -303,9 +268,8 @@ function constraint_ohms_yt_from_on_off{T <: AbstractWRForm}(pm::GenericPowerMod
     wr = pm.var[:wr][i]
     wi = pm.var[:wi][i]
 
-    c1 = @constraint(pm.model, p_fr ==        g/tm*w_fr + (-g*tr+b*ti)/tm*(wr) + (-b*tr-g*ti)/tm*( wi) )
-    c2 = @constraint(pm.model, q_fr == -(b+c/2)/tm*w_fr - (-b*tr-g*ti)/tm*(wr) + (-g*tr+b*ti)/tm*( wi) )
-    return Set([c1, c2])
+    @constraint(pm.model, p_fr ==        g/tm*w_fr + (-g*tr+b*ti)/tm*(wr) + (-b*tr-g*ti)/tm*( wi) )
+    @constraint(pm.model, q_fr == -(b+c/2)/tm*w_fr - (-b*tr-g*ti)/tm*(wr) + (-g*tr+b*ti)/tm*( wi) )
 end
 
 """
@@ -323,9 +287,8 @@ function constraint_ohms_yt_to_on_off{T <: AbstractWRForm}(pm::GenericPowerModel
     wr = pm.var[:wr][i]
     wi = pm.var[:wi][i]
 
-    c1 = @constraint(pm.model, p_to ==        g*w_to + (-g*tr-b*ti)/tm*(wr) + (-b*tr+g*ti)/tm*(-wi) )
-    c2 = @constraint(pm.model, q_to == -(b+c/2)*w_to - (-b*tr+g*ti)/tm*(wr) + (-g*tr-b*ti)/tm*(-wi) )
-    return Set([c1, c2])
+    @constraint(pm.model, p_to ==        g*w_to + (-g*tr-b*ti)/tm*(wr) + (-b*tr+g*ti)/tm*(-wi) )
+    @constraint(pm.model, q_to == -(b+c/2)*w_to - (-b*tr+g*ti)/tm*(wr) + (-g*tr-b*ti)/tm*(-wi) )
 end
 
 "`angmin*wr[i] <= wi[i] <= angmax*wr[i]`"
@@ -333,9 +296,8 @@ function constraint_phase_angle_difference_on_off{T <: AbstractWRForm}(pm::Gener
     wr = pm.var[:wr][i]
     wi = pm.var[:wi][i]
 
-    c1 = @constraint(pm.model, wi <= tan(angmax)*wr)
-    c2 = @constraint(pm.model, wi >= tan(angmin)*wr)
-    return Set([c1, c2])
+    @constraint(pm.model, wi <= tan(angmax)*wr)
+    @constraint(pm.model, wi >= tan(angmin)*wr)
 end
 
 "`angmin*wr_ne[i] <= wi_ne[i] <= angmax*wr_ne[i]`"
@@ -343,9 +305,8 @@ function constraint_phase_angle_difference_ne{T <: AbstractWRForm}(pm::GenericPo
     wr = pm.var[:wr_ne][i]
     wi = pm.var[:wi_ne][i]
 
-    c1 = @constraint(pm.model, wi <= tan(angmax)*wr)
-    c2 = @constraint(pm.model, wi >= tan(angmin)*wr)
-    return Set([c1, c2])
+    @constraint(pm.model, wi <= tan(angmax)*wr)
+    @constraint(pm.model, wi >= tan(angmin)*wr)
 end
 
 ""
@@ -366,8 +327,6 @@ function variable_voltage_magnitude_sqr_from_ne{T <: AbstractWRForm}(pm::Generic
         upperbound = buses[branches[i]["f_bus"]]["vmax"]^2,
         start = getstart(pm.ref[:bus], i, "w_from_start", 1.001)
     )
-
-    return pm.var[:w_from_ne]
 end
 
 ""
@@ -381,8 +340,6 @@ function variable_voltage_magnitude_sqr_to_ne{T <: AbstractWRForm}(pm::GenericPo
         upperbound = buses[branches[i]["t_bus"]]["vmax"]^2,
         start = getstart(pm.ref[:bus], i, "w_to", 1.001)
     )
-
-    return pm.var[:w_to_ne]
 end
 
 ""
@@ -403,8 +360,6 @@ function variable_voltage_product_ne{T <: AbstractWRForm}(pm::GenericPowerModel{
         upperbound = max(0, wi_max[bi_bp[b]]), 
         start = getstart(pm.ref[:ne_buspairs], bi_bp[b], "wi_start")
     )
-
-    return pm.var[:wr_ne], pm.var[:wi_ne]
 end
 
 ""
@@ -429,7 +384,6 @@ function variable_phase_angle_difference{T}(pm::GenericPowerModel{T})
         upperbound = pm.ref[:buspairs][bp]["angmax"], 
         start = getstart(pm.ref[:buspairs], bp, "td_start")
     )
-    return pm.var[:td]
 end
 
 "Creates the voltage magnitude product variables"
@@ -441,7 +395,6 @@ function variable_voltage_magnitude_product{T}(pm::GenericPowerModel{T})
         upperbound = buspairs[bp]["v_from_max"]*buspairs[bp]["v_to_max"],
         start = getstart(pm.ref[:buspairs], bp, "vv_start", 1.0)
     )
-    return pm.var[:vv]
 end
 
 ""
@@ -470,7 +423,6 @@ function variable_cosine{T}(pm::GenericPowerModel{T})
         upperbound = cos_max[bp],
         start = getstart(pm.ref[:buspairs], bp, "cs_start", 1.0)
     )
-    return pm.var[:cs]
 end
 
 ""
@@ -481,7 +433,6 @@ function variable_sine(pm::GenericPowerModel)
         upperbound = sin(pm.ref[:buspairs][bp]["angmax"]), 
         start = getstart(pm.ref[:buspairs], bp, "si_start")
     )
-    return pm.var[:si]
 end
 
 ""
@@ -493,7 +444,6 @@ function variable_current_magnitude_sqr{T}(pm::GenericPowerModel{T})
         upperbound = (buspairs[bp]["rate_a"]*buspairs[bp]["tap"]/buspairs[bp]["v_from_min"])^2,
         start = getstart(pm.ref[:buspairs], bp, "cm_start")
     )
-    return pm.var[:cm]
 end
 
 ""
@@ -525,24 +475,20 @@ function constraint_voltage(pm::QCWRPowerModel)
     wr = pm.var[:wr]
     wi = pm.var[:wi]
 
-    const_set = Set()
     for (i,b) in pm.ref[:bus]
-        cs1 = relaxation_sqr(pm.model, v[i], w[i])
-        const_set = union(const_set, cs1)
+        relaxation_sqr(pm.model, v[i], w[i])
     end
 
     for bp in keys(pm.ref[:buspairs])
         i,j = bp
-        c1 = @constraint(pm.model, t[i] - t[j] == td[bp])
-        push!(const_set, c1)
+        @constraint(pm.model, t[i] - t[j] == td[bp])
 
-        cs1 = relaxation_sin(pm.model, td[bp], si[bp])
-        cs2 = relaxation_cos(pm.model, td[bp], cs[bp])
-        cs3 = relaxation_product(pm.model, v[i], v[j], vv[bp])
-        cs4 = relaxation_product(pm.model, vv[bp], cs[bp], wr[bp])
-        cs5 = relaxation_product(pm.model, vv[bp], si[bp], wi[bp])
+        relaxation_sin(pm.model, td[bp], si[bp])
+        relaxation_cos(pm.model, td[bp], cs[bp])
+        relaxation_product(pm.model, v[i], v[j], vv[bp])
+        relaxation_product(pm.model, vv[bp], cs[bp], wr[bp])
+        relaxation_product(pm.model, vv[bp], si[bp], wi[bp])
 
-        const_set = union(const_set, cs1, cs2, cs3, cs4, cs5)
         # this constraint is redudant and useful for debugging
         #relaxation_complex_product(pm.model, w[i], w[j], wr[bp], wi[bp])
    end
@@ -553,13 +499,11 @@ function constraint_voltage(pm::QCWRPowerModel)
 
         # to prevent this constraint from being posted on multiple parallel lines
         if buspair["line"] == i
-            cs1 = constraint_power_magnitude_sqr(pm, branch)
-            cs2 = constraint_power_magnitude_link(pm, branch)
-            const_set = union(const_set, cs1, cs2)
+            constraint_power_magnitude_sqr(pm, branch)
+            constraint_power_magnitude_link(pm, branch)
         end
     end
 
-    return const_set
 end
 
 "`p[f_idx]^2 + q[f_idx]^2 <= w[f_bus]/tm*cm[f_bus,t_bus]`"
@@ -569,8 +513,7 @@ function constraint_power_magnitude_sqr(pm::QCWRPowerModel, f_bus, t_bus, arc_fr
     q_fr = pm.var[:q][arc_from]
     cm = pm.var[:cm][(f_bus, t_bus)]
 
-    c = @constraint(pm.model, p_fr^2 + q_fr^2 <= w_i/tm*cm)
-    return Set([c])
+    @constraint(pm.model, p_fr^2 + q_fr^2 <= w_i/tm*cm)
 end
 
 "`cm[f_bus,t_bus] == (g^2 + b^2)*(w[f_bus]/tm + w[t_bus] - 2*(tr*wr[f_bus,t_bus] + ti*wi[f_bus,t_bus])/tm) - c*q[f_idx] - ((c/2)/tm)^2*w[f_bus]`"
@@ -582,8 +525,7 @@ function constraint_power_magnitude_link(pm::QCWRPowerModel, f_bus, t_bus, arc_f
     wi = pm.var[:wi][(f_bus, t_bus)]
     cm = pm.var[:cm][(f_bus, t_bus)]
 
-    c = @constraint(pm.model, cm == (g^2 + b^2)*(w_fr/tm + w_to - 2*(tr*wr + ti*wi)/tm) - c*q_fr - ((c/2)/tm)^2*w_fr)
-    return Set([c])
+    @constraint(pm.model, cm == (g^2 + b^2)*(w_fr/tm + w_to - 2*(tr*wr + ti*wi)/tm) - c*q_fr - ((c/2)/tm)^2*w_fr)
 end
 
 "`t[ref_bus] == 0`"
@@ -607,12 +549,10 @@ function constraint_phase_angle_difference(pm::QCWRPowerModel, f_bus, t_bus, ang
     wr = pm.var[:wr][(f_bus, t_bus)]
     wi = pm.var[:wi][(f_bus, t_bus)]
 
-    c1 = @constraint(pm.model, wi <= tan(angmax)*wr)
-    c2 = @constraint(pm.model, wi >= tan(angmin)*wr)
+    @constraint(pm.model, wi <= tan(angmax)*wr)
+    @constraint(pm.model, wi >= tan(angmin)*wr)
 
-    c3 = cut_complex_product_and_angle_difference(pm.model, w_fr, w_to, wr, wi, angmin, angmax)
-
-    return Set([c1, c2, c3])
+    cut_complex_product_and_angle_difference(pm.model, w_fr, w_to, wr, wi, angmin, angmax)
 end
 
 ""
@@ -652,7 +592,6 @@ function variable_phase_angle_difference_on_off{T}(pm::GenericPowerModel{T})
         upperbound = max(0, pm.ref[:branch][l]["angmax"]),
         start = getstart(pm.ref[:branch], l, "td_start")
     )
-    return pm.var[:td]
 end
 
 ""
@@ -666,8 +605,6 @@ function variable_voltage_magnitude_product_on_off{T}(pm::GenericPowerModel{T})
         upperbound = max(0, vv_max[l]),
         start = getstart(pm.ref[:branch], l, "vv_start", 1.0)
     )
-
-    return pm.var[:vv]
 end
 
 
@@ -697,8 +634,6 @@ function variable_cosine_on_off{T}(pm::GenericPowerModel{T})
         upperbound = max(0, cos_max[l]), 
         start = getstart(pm.ref[:branch], l, "cs_start", 1.0)
     )
-
-    return pm.var[:cs]
 end
 
 ""
@@ -709,7 +644,6 @@ function variable_sine_on_off(pm::GenericPowerModel)
         upperbound = max(0, sin(pm.ref[:branch][l]["angmax"])),
         start = getstart(pm.ref[:branch], l, "si_start")
     )
-    return pm.var[:si]
 end
 
 
@@ -724,8 +658,6 @@ function variable_current_magnitude_sqr_on_off{T}(pm::GenericPowerModel{T})
         upperbound = cm_max[l],
         start = getstart(pm.ref[:branch], l, "cm_start")
     )
-
-    return pm.var[:cm]
 end
 
 
@@ -754,52 +686,42 @@ function constraint_voltage_on_off(pm::QCWRPowerModel)
     td_ub = pm.ref[:off_angmax]
     td_max = max(abs(td_lb), abs(td_ub))
 
-    #cs = Set()
     for (i,b) in pm.ref[:bus]
-        cs1 = relaxation_sqr(pm.model, v[i], w[i])
-        #cs = union(cs, cs1)
+        relaxation_sqr(pm.model, v[i], w[i])
     end
 
-    cs1 = constraint_voltage_magnitude_from_on_off(pm) # bounds on v_from
-    cs2 = constraint_voltage_magnitude_to_on_off(pm) # bounds on v_to
-    cs3 = constraint_voltage_magnitude_sqr_from_on_off(pm) # bounds on w_from
-    cs4 = constraint_voltage_magnitude_sqr_to_on_off(pm) # bounds on w_to
-    cs5 = constraint_voltage_product_on_off(pm) # bounds on wr, wi
-    #cs = union(cs, cs1, cs2, cs3, cs4, cs5)
+    constraint_voltage_magnitude_from_on_off(pm) # bounds on v_from
+    constraint_voltage_magnitude_to_on_off(pm) # bounds on v_to
+    constraint_voltage_magnitude_sqr_from_on_off(pm) # bounds on w_from
+    constraint_voltage_magnitude_sqr_to_on_off(pm) # bounds on w_to
+    constraint_voltage_product_on_off(pm) # bounds on wr, wi
 
     for (l,branch) in pm.ref[:branch]
         i = branch["f_bus"]
         j = branch["t_bus"]
 
-        c1 = @constraint(pm.model, t[i] - t[j] >= td[l] + td_lb*(1-z[l]))
-        c2 = @constraint(pm.model, t[i] - t[j] <= td[l] + td_ub*(1-z[l]))
-        #cs = union(cs, Set([c1, c2]))
+        @constraint(pm.model, t[i] - t[j] >= td[l] + td_lb*(1-z[l]))
+        @constraint(pm.model, t[i] - t[j] <= td[l] + td_ub*(1-z[l]))
 
-        cs1 = relaxation_sin_on_off(pm.model, td[l], si[l], z[l], td_max)
-        cs2 = relaxation_cos_on_off(pm.model, td[l], cs[l], z[l], td_max)
-        cs3 = relaxation_product_on_off(pm.model, v_from[i], v_to[j], vv[l], z[l])
-        cs4 = relaxation_product_on_off(pm.model, vv[l], cs[l], wr[l], z[l])
-        cs5 = relaxation_product_on_off(pm.model, vv[l], si[l], wi[l], z[l])
-        #const_set = union(const_set, cs1, cs2, cs3, cs4, cs5)
+        relaxation_sin_on_off(pm.model, td[l], si[l], z[l], td_max)
+        relaxation_cos_on_off(pm.model, td[l], cs[l], z[l], td_max)
+        relaxation_product_on_off(pm.model, v_from[i], v_to[j], vv[l], z[l])
+        relaxation_product_on_off(pm.model, vv[l], cs[l], wr[l], z[l])
+        relaxation_product_on_off(pm.model, vv[l], si[l], wi[l], z[l])
 
         # this constraint is redudant and useful for debugging
         #relaxation_complex_product(pm.model, w[i], w[j], wr[l], wi[l])
 
-        #cs4 = relaxation_complex_product_on_off(pm.model, w[i], w[j], wr[l], wi[l], z[l])
-        cs6 = relaxation_equality_on_off(pm.model, v[i], v_from[l], z[l])
-        cs7 = relaxation_equality_on_off(pm.model, v[j], v_to[l], z[l])
-        cs8 = relaxation_equality_on_off(pm.model, w[i], w_from[l], z[l])
-        cs9 = relaxation_equality_on_off(pm.model, w[j], w_to[l], z[l])
-        #cs = union(cs, cs6, cs7, cs8, cs9)
+        relaxation_equality_on_off(pm.model, v[i], v_from[l], z[l])
+        relaxation_equality_on_off(pm.model, v[j], v_to[l], z[l])
+        relaxation_equality_on_off(pm.model, w[i], w_from[l], z[l])
+        relaxation_equality_on_off(pm.model, w[j], w_to[l], z[l])
 
         # to prevent this constraint from being posted on multiple parallel lines
         # TODO needs on/off variant
-        cs1 = constraint_power_magnitude_sqr_on_off(pm, branch)
-        cs2 = constraint_power_magnitude_link_on_off(pm, branch) # different index set
-        #cs = union(cs, cs1, cs2)
+        constraint_power_magnitude_sqr_on_off(pm, branch)
+        constraint_power_magnitude_link_on_off(pm, branch) # different index set
     end
-
-    return Set()
 end
 
 
@@ -816,11 +738,9 @@ function constraint_power_magnitude_sqr_on_off(pm::QCWRPowerModel, i, f_bus, arc
     cm_ub = getupperbound(cm)
     z_ub = getupperbound(z)
 
-    c1 = @constraint(pm.model, p_fr^2 + q_fr^2 <= w*cm*z_ub/tm)
-    c2 = @constraint(pm.model, p_fr^2 + q_fr^2 <= w_ub*cm*z/tm)
-    c3 = @constraint(pm.model, p_fr^2 + q_fr^2 <= w*cm_ub*z/tm)
-
-    return Set([c1, c2, c3])
+    @constraint(pm.model, p_fr^2 + q_fr^2 <= w*cm*z_ub/tm)
+    @constraint(pm.model, p_fr^2 + q_fr^2 <= w_ub*cm*z/tm)
+    @constraint(pm.model, p_fr^2 + q_fr^2 <= w*cm_ub*z/tm)
 end
 
 "`cm[f_bus,t_bus] == (g^2 + b^2)*(w[f_bus]/tm + w[t_bus] - 2*(tr*wr[f_bus,t_bus] + ti*wi[f_bus,t_bus])/tm) - c*q[f_idx] - ((c/2)/tm)^2*w[f_bus]`"
@@ -832,6 +752,5 @@ function constraint_power_magnitude_link_on_off(pm::QCWRPowerModel, i, arc_from,
     wi = pm.var[:wi][i]
     cm = pm.var[:cm][i]
 
-    c = @constraint(pm.model, cm == (g^2 + b^2)*(w_fr/tm + w_to - 2*(tr*wr + ti*wi)/tm) - c*q_fr - ((c/2)/tm)^2*w_fr)
-    return Set([c])
+    @constraint(pm.model, cm == (g^2 + b^2)*(w_fr/tm + w_to - 2*(tr*wr + ti*wi)/tm) - c*q_fr - ((c/2)/tm)^2*w_fr)
 end